update frame content
