+++ network/src/commonMain/kotlin/com/wire/kalium/network/KaliumKtorCustomLogging.kt
@@ -1,7 +1,8 @@
 package com.wire.kalium.network
 
-import com.wire.kalium.logger.obfuscateId
-import com.wire.kalium.logger.obfuscateUrlPath
+import com.wire.kalium.network.utils.obfuscateAndLogMessage
+import com.wire.kalium.network.utils.obfuscatePath
+import com.wire.kalium.network.utils.sensitiveJsonKeys
 import io.ktor.client.HttpClient
 import io.ktor.client.HttpClientConfig
 import io.ktor.client.plugins.HttpClientPlugin
@@ -249,62 +250,6 @@ public class KaliumKtorCustomLogging private constructor(
         }
     }
 }
-
-@Suppress("TooGenericExceptionCaught")
-private fun obfuscateAndLogMessage(text: String) {
-    try {
-        val obj = (Json.decodeFromString(text) as JsonElement)
-        if (obj.jsonArray.size > 0) {
-            obj.jsonArray.map {
-                logObfuscatedJsonElement(it)
-            }
-        } else {
-            logObfuscatedJsonElement(obj)
-        }
-        obj.toString()
-    } catch (e: Exception) {
-        "error the body content while logging "
-    }
-}
-
-fun logObfuscatedJsonElement(obj: JsonElement) {
-    obj.jsonObject.entries.toMutableSet().map {
-        when {
-            sensitiveJsonKeys.contains(it.key.lowercase()) -> {
-                kaliumLogger.v("${it.key} : ******")
-            }
-            sensitiveJsonIdKeys.contains(it.key.lowercase()) -> {
-                kaliumLogger.v("${it.key} : ${it.value.toString().obfuscateId()}")
-            }
-            sensitiveJsonObjects.contains(it.key.lowercase()) -> {
-                logObfuscatedJsonElement(it.value)
-            }
-            else -> {
-                kaliumLogger.v("${it.key} : ${it.value}")
-            }
-        }
-    }
-}
-
-private fun obfuscatePath(url: Url): String {
-    var requestToLog = url.host
-    if (url.pathSegments.isNotEmpty()) {
-        url.pathSegments.map {
-            requestToLog += "/${it.obfuscateUrlPath()}"
-        }
-    }
-    url.parameters.entries().map {
-        if (it.value.isNotEmpty()) {
-            requestToLog += "/${it.key} = ${it.value[0].obfuscateUrlPath()}"
-        }
-    }
-    return requestToLog
-}
-
-
-private val sensitiveJsonKeys by lazy { listOf("password", "authorization", "set-cookie","Location", "x-amz-meta-user") }
-private val sensitiveJsonIdKeys by lazy { listOf("conversation", "id", "user", "team") }
-private val sensitiveJsonObjects by lazy { listOf("qualified_id") }
 
 /**
  * Configure and install [Logging] in [HttpClient].
